- fixed size_t<->int conversion warnings - fixed issue with slow start on multiple monitors (now searching for gold sheep just once, instead of on each AddDisplay)

+++ client_generic/ContentDownloader/SheepGenerator.cpp
@@ -161,7 +161,7 @@ void	SheepGenerator::setURL( const char *url )
  	boost::mutex::scoped_lock lockthis( s_GeneratorMutex );
 
 	SAFE_DELETE_ARRAY( fURL );
-	int len = strlen( url );
+	size_t len = strlen( url );
 	fURL = new char[ len + 1 ];
 	strncpy( fURL, url, len + 1 );
 }
